Break out utility functions into a separate file
