+++ tracpro/polls/models.py
@@ -2,15 +2,11 @@ from __future__ import absolute_import, unicode_literals
 
 from collections import Counter, defaultdict, OrderedDict
 from decimal import Decimal, InvalidOperation
-import math
 import operator
-import re
 
 from dateutil.relativedelta import relativedelta
 from enum import Enum
-import pycountry
 import pytz
-import stop_words
 
 from django.conf import settings
 from django.core.cache import cache
@@ -23,7 +19,9 @@ from django.utils.translation import ugettext_lazy as _
 from dash.utils import get_cacheable, get_month_range
 
 from tracpro.contacts.models import Contact
+
 from .tasks import pollrun_start
+from .utils import auto_range_categories, extract_words
 
 
 QUESTION_TYPE_OPEN = 'O'
@@ -579,46 +577,3 @@ class Answer(models.Model):
             category_counts[label] += 1
 
         return category_counts
-
-
-def auto_range_categories(value_min, value_max):
-    """
-    Tries to pick sensible numerical range categories for the given range of values
-    """
-    value_range = value_max - value_min
-    if value_range > 1:
-        category_range = int(math.pow(10, math.ceil(math.log10(value_range))))  # 10, 100, etc
-
-        if value_range < category_range / 2:
-            category_range /= 2  # 5, 10, 50, 100 etc
-    else:
-        category_range = 5
-
-    category_step = category_range / 5  # aim for 5 categories
-
-    category_min = value_min - (value_min % category_step)
-
-    category_max = category_min + category_step * 5
-
-    # may need an extra category to hold max value
-    while category_max <= value_max:
-        category_max += category_step
-
-    return category_min, category_max, category_step
-
-
-def extract_words(text, language):
-    """
-    Extracts significant words from the given text (i.e. words we want to include in a word cloud)
-    """
-    ignore_words = []
-    if language:
-        code = pycountry.languages.get(bibliographic=language).alpha2
-        try:
-            ignore_words = stop_words.get_stop_words(code)
-        except stop_words.StopWordError:
-            pass
-
-    words = re.split(r"[^\w'-]", text.lower(), flags=re.UNICODE)
-    ignore_words = ignore_words
-    return [w for w in words if w not in ignore_words and len(w) > 1]
